Move implementations into AccessLists
- Related to #2096.
- Resolves #2097.
